+++ vscode-extension/extension.ts
@@ -1,9 +1,12 @@
 import * as vscode from 'vscode'
 import * as glob from 'glob'
+import * as Markdown from 'markdown-it'
 import { BuildModuleCommand } from './Commands/BuildModuleCommand'
 import { CreateModuleJsonCommand } from './Commands/CreateModuleJsonCommand'
 import { ExportToPdfCommand } from './Commands/ExportToPdfCommand'
 import { MarkdownToggler } from './Commands/MarkdownToggler'
+import { ModuleProjectProvider } from './TreeViewProviders/ModuleProjectProvider'
+import { MarkdownRenderer } from '../shared/MarkdownRenderer'
 
 const buildModuleCommand = new BuildModuleCommand()
 const createModuleJsonCommand = new CreateModuleJsonCommand()
@@ -12,16 +15,35 @@ const exportToPdfCommand = new ExportToPdfCommand()
 export function activate(context: vscode.ExtensionContext) {
   console.log('EncounterPlus Markdown Extension loaded.')
 
+  if (vscode.workspace.rootPath === undefined) {
+    return
+  }
+
   glob(vscode.workspace.rootPath + '/**/*.md', {}, (error, matches) => {
     if (matches && matches.length > 0) {
       vscode.commands.executeCommand('setContext', 'projectHasMarkdown', true)
     }
   })
 
+  let moduleProjectProvider = new ModuleProjectProvider(vscode.workspace.rootPath)
+  vscode.window.registerTreeDataProvider('encounter-plus-modules', moduleProjectProvider)
+  context.subscriptions.push(
+    vscode.commands.registerCommand('encounterPlusMarkdown.refreshModules', () => moduleProjectProvider.refresh())
+  )
+
+  context.subscriptions.push(
+    vscode.commands.registerCommand('encounterPlusMarkdown.openPage', async (pagePath) => {
+      let uri = vscode.Uri.file(pagePath)
+      let doc = await vscode.workspace.openTextDocument(uri)
+      vscode.window.showTextDocument(doc, vscode.ViewColumn.One)
+      //vscode.commands.executeCommand
+    })
+  )
+
   let markdownToggler = new MarkdownToggler()
   let toggles = markdownToggler.toggleDictionary
   for (const command in toggles) {
-    vscode.commands.registerTextEditorCommand(command, textEditor => {
+    vscode.commands.registerTextEditorCommand(command, (textEditor) => {
       markdownToggler.toggleFormat(textEditor, command)
     })
   }
@@ -32,30 +54,21 @@ export function activate(context: vscode.ExtensionContext) {
     })
   )
 
-  context.subscriptions.push(    
+  context.subscriptions.push(
     vscode.commands.registerCommand('encounterPlusMarkdown.createModuleJson', () => {
       createModuleJsonCommand.startCommand()
     })
   )
 
-  context.subscriptions.push(    
+  context.subscriptions.push(
     vscode.commands.registerCommand('encounterPlusMarkdown.exportModuleToPDF', () => {
       exportToPdfCommand.startCommand()
     })
   )
 
   return {
-    extendMarkdownIt(md: any) {
-      return md
-        .use(require('markdown-it-anchor'))
-        .use(require('markdown-it-attrs'))
-        .use(require('markdown-it-decorate'))
-        .use(require('markdown-it-imsize'), { autofill: true })
-        .use(require('markdown-it-mark'))
-        .use(require('markdown-it-multimd-table'))        
-        .use(require('markdown-it-sub'))
-        .use(require('markdown-it-sup'))
-        .use(require('markdown-it-underline'))
+    extendMarkdownIt(md: Markdown) {
+      return MarkdownRenderer.getRenderer()
     },
   }
 }